final files added

+++ Contact Management - Project/mycontacts-backend/controllers/contactController.js
@@ -2,10 +2,10 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json({
     contacts,
   });
@@ -13,7 +13,7 @@ const getContacts = asyncHandler(async (req, res) => {
 
 //@desc Get contact by id
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
   const contacts = await Contact.findById(req.params.id);
@@ -28,19 +28,20 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Create new contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
   console.log(req.body);
   const { name, email, phone } = req.body;
   if (!name || !email || !phone) {
-    res.status(404);
-    throw new Error("All fields are mandetory");
+    res.status(400); // Change status to 400
+    throw new Error("All fields are mandatory"); // Correct the error message
   }
   const contact = await Contact.create({
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(201).json({
     contact,
@@ -49,36 +50,44 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Update contact by id
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.findById(req.params.id);
-  if (!contacts) {
+  const contact = await Contact.findById(req.params.id);
+  if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
-  const updatedContact=await Contact.findByIdAndUpdate(
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User doesn't have permission to update other user's contacts");
+  }
+  const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
-    {new: true}
-  )
-  res.status(201).json({
+    { new: true }
+  );
+  res.status(200).json({
     updatedContact,
   });
 });
 
 //@desc Delete contact by id
 //@route delete /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
-    res.status(404).json({ message: "Contact not found" });
-  } else {
-    await Contact.deleteOne({ _id: req.params.id });
-    res.status(200).json({ message: "Contact deleted" });
+    res.status(404);
+    throw new Error("Contact not found");
   }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User dont have permission to delete other user's contact's ");
+  }
+  await Contact.deleteOne({ _id: req.params.id });
+  res.status(200).json({ message: "Contact deleted" });
 });
 
 module.exports = {
